pyproject.toml
[project]
name = "owl-bot"
version = "0.1.0"
description = "🦉 A witty Discord bot: pronunciation, definitions, translation, transcription, and rating."
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "discord.py>=2.4.0",
    "openai>=1.40.0",
    "gTTS>=2.5.1",
    "whisper @ git+https://github.com/openai/whisper.git",
    "fasttext-wheel>=0.9.2",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.5"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
